checkpoint: renames "useStatefulAttemptThatEventually.ts" to "instance.ts"
