+++ src/library/vue/ProgressiveRef/exports.toolbox.ts
@@ -1,1 +1,1 @@
-export * from './useStatefulAttemptThatEventually';
+export * from './instance';
